moustache_backend/requirements.txt
django>=4.0.0
geopy>=2.3.0
requests>=2.26.0
python-dotenv>=0.19.0
openai>=0.27.0
